+++ CDP4Composition.Tests/CDP4Composition.Tests.csproj
@@ -126,12 +126,4 @@
   <ItemGroup>
     <EmbeddedResource Include="Resources\RibbonXml\testribbon.xml" />
   </ItemGroup>
-  <ItemGroup>
-    <None Update="ViewModels\PluginMockData\CDP4BasicRdl.cdp4ck">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Update="ViewModels\PluginMockData\CDP4BasicRdl.plugin.manifest">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
 </Project>
