+++ CDP4IME.Tests/CDP4IME-CE.Tests.csproj
@@ -9,6 +9,7 @@
     <Product>CDP4IME.Tests</Product>
     <Description>CDP4IME Unit Tests</Description>
     <Copyright>Copyright © RHEA System S.A.</Copyright>
+    <LangVersion>latest</LangVersion>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugType>full</DebugType>
@@ -58,6 +59,7 @@
     <PackageReference Include="System.Xml.XmlSerializer" Version="4.3.0" />
   </ItemGroup>
   <ItemGroup>
+    <Reference Include="PresentationCore" />
     <Reference Include="System.ComponentModel.Composition" />
     <Reference Include="System.Configuration" />
     <Reference Include="System.IO.Compression" />
@@ -76,5 +78,13 @@
     <Content Include="Test.txt">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <None Update="ViewModels\PluginMockData\CDP4BasicRdl.cdp4ck">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="ViewModels\PluginMockData\CDP4BasicRdl.plugin.manifest">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 </Project>
