+++ CDP4IME.Tests/ViewModels/PluginMockData/CDP4BasicRdl.plugin.manifest
