+++ CDP4IME/Views/PluginInstaller.xaml
@@ -1,11 +1,11 @@
-﻿<Window x:Class="CDP4Composition.Views.PluginInstaller"
+﻿<Window x:Class="CDP4IME.Views.PluginInstaller"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:CDP4Composition.Views"
+        xmlns:local="clr-namespace:CDP4IME.Views"
         xmlns:dx="http://schemas.devexpress.com/winfx/2008/xaml/core"
-        xmlns:behaviors="clr-namespace:CDP4Composition.Behaviors"
+        xmlns:behaviors="clr-namespace:CDP4IME.Behaviors"
         xmlns:dxmvvm="http://schemas.devexpress.com/winfx/2008/xaml/mvvm"
         mc:Ignorable="d"
         Title="CDP4-Plugin-Installer" Height="600" Width="600" ResizeMode="NoResize" BorderThickness="0" Background="White">
