+++ CDP4IME/Views/PluginRow.xaml
@@ -1,10 +1,10 @@
-﻿<UserControl x:Class="CDP4Composition.Views.PluginRow"
+﻿<UserControl x:Class="CDP4IME.Views.PluginRow"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:CDP4Composition.Views"
-             xmlns:converters="clr-namespace:CDP4Composition.Converters"
+             xmlns:local="clr-namespace:CDP4IME.Views"
+             xmlns:converters="clr-namespace:CDP4Composition.Converters;assembly=CDP4Composition"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800" Margin="5" HorizontalAlignment="Stretch" HorizontalContentAlignment="Stretch">
     <UserControl.Resources>
@@ -30,7 +30,7 @@
             <Label Content="{Binding Version}" Margin="10,0,0,0" VerticalContentAlignment="Bottom" FontSize="14" HorizontalAlignment="Right"></Label>
         </StackPanel>
         <StackPanel Orientation="Horizontal" VerticalAlignment="Center" Grid.Row="0" Grid.Column="1" HorizontalAlignment="Right">
-            <Image Margin="3" VerticalAlignment="Center" Source="../Resources/Images/rhea.png" Height="18"></Image>
+            <Image Margin="3" VerticalAlignment="Center" Source="../Resources/rhea.png" Height="18"></Image>
             <Label VerticalAlignment="Center" Content="{Binding Author}" HorizontalAlignment="Right"></Label>
         </StackPanel>
         <TextBlock HorizontalAlignment="Center" Grid.Row="1" Grid.Column="1" VerticalAlignment="Bottom">
